constructed getCharacterChoice and playRound
The game works for one round currently, though text is not
yet converted to first-character-capitalization. The playRound function
works as intended, but its code is quite bloated and could likely use
a switch instead for greater efficiency. Moreover, the game still
requires a game function as well as a games won tally for best of 5 play